requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
exceptiongroup==1.2.2
fonttools==4.56.0
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
kiwisolver==1.4.8
matplotlib==3.10.1
mysqlclient==2.2.7
numpy==2.2.4
openai==1.74.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
